Fix headers check for HTML content-type
For normal requests, headers are using the `Rack::Headers` where keys are case-insensitive.

In `ActionDispatch::ShowExceptions`, Rails uses a Hash instead which is case-sensitive. It uses the `Rack::CONTENT_TYPE` constant that is a down-case version of the header key.

In this commit, I changed the "Content-Type" key to "content-type" so that live reloading still works in the Rails error page.